feat(ender): moved some code to ee + fixed mem leak
